added stuff

+++ gui/AlbumWidget.py
@@ -1,38 +1,79 @@
 from PySide2.QtWidgets import QGroupBox
+from PySide2.QtWidgets import QTabWidget
 from PySide2.QtWidgets import QVBoxLayout
 from PySide2.QtWidgets import QHBoxLayout
 from PySide2.QtWidgets import QGridLayout
 from PySide2.QtWidgets import QFormLayout
+from PySide2.QtWidgets import QWidget
 from PySide2.QtWidgets import QLabel
 from PySide2.QtWidgets import QComboBox
 from PySide2.QtWidgets import QPushButton
 from PySide2.QtWidgets import QSlider
 from PySide2.QtWidgets import QLineEdit
+from PySide2.QtWidgets import QMessageBox
+from PySide2.QtWidgets import QListWidget
+from PySide2.QtWidgets import QListWidgetItem
 from PySide2.QtGui import QPixmap
 from PySide2.QtGui import QIcon
 from PySide2.QtCore import QSize
 from PySide2.QtCore import Qt
 from gui.AddSongDialog import AddSongDialog
 from gui.AlbumSortView import AlbumSortView
+from gui.AlbumSongListItem import AlbumSongListItem
+from gui.EditAlbumDialog import EditAlbumDialog
 
-class AlbumWidget(QGroupBox):
+class AlbumWidget(QTabWidget):
     def __init__(self, db, album, parent):
         super().__init__(parent)
         self.db = db
         self.album = album
         self.parent = parent
-        self.setTitle(self.album.title)
+        self.overview = QWidget()
+        self.songs = QWidget()
+        self.init_overview()
+        self.init_songs()
+        self.addTab(self.overview, "Overview")
+        self.addTab(self.songs, "Songs")
+
+
+    def init_overview(self):
         self.layout = QVBoxLayout()
-        self.setLayout(self.layout)
-        self.setFixedSize(QSize(350, 450))
-        self.init_remove()
+        self.overview.setLayout(self.layout)
+        self.init_top()
         self.init_cover_area()
         self.init_data_area()
 
-    def init_remove(self):
+    def init_songs(self):
+        self.songs.deleteLater()
+        self.songs = QWidget()
+        self.song_layout = QVBoxLayout()
+        self.song_list = QListWidget()
+
+        for song in self.album.songs:
+            item = QListWidgetItem(self.song_list)
+            self.song_list.addItem(item)
+            song_item = AlbumSongListItem(self.db, self.album, song)
+            item.setSizeHint(song_item.sizeHint())
+            self.song_list.setItemWidget(item, song_item)
+        
+        self.song_layout.addWidget(self.song_list)
+        self.songs.setLayout(self.song_layout)
+        
+
+
+    def init_top(self):
+        top_layout = QHBoxLayout()
+        title_edit = QLineEdit(self.album.title)
+        title_edit.setReadOnly(True)
+        edit_button = QPushButton("Edit")
+        edit_button.clicked.connect(self.edit)
         delete_button = QPushButton("Delete")
         delete_button.clicked.connect(self.delete)
-        self.layout.addWidget(delete_button)
+        top_layout.addWidget(title_edit)
+        top_layout.addStretch()
+        top_layout.addWidget(edit_button)
+        top_layout.addWidget(delete_button)
+        self.layout.addLayout(top_layout)
 
     def init_cover_area(self):
         cover_area = QGroupBox()
@@ -119,8 +160,21 @@ class AlbumWidget(QGroupBox):
         self.rating_edit.setText(f"{self.album.rating}/10")
 
     def delete(self):
-        self.db.delete_album(self.album)
-        self.db.album_view.refresh()        
+        msg_box = QMessageBox()
+        msg_box.setWindowTitle("Delete Album?")
+        msg_box.setText("Yeet?")
+        msg_box.setStandardButtons(QMessageBox.Yes | QMessageBox.Cancel)
+        msg_box.setDefaultButton(QMessageBox.Cancel)
+        ret = msg_box.exec_()
+
+        if ret == QMessageBox.Yes:    
+            self.db.delete_album(self.album)
+            self.db.album_view.refresh()
+            self.db.global_rating.refresh()
+            self.db.user_rating.refresh()       
+
+    def edit(self):
+        view = EditAlbumDialog(self.db, self.album, self) 
 
     def open_album_sort_view(self):
         view = AlbumSortView(self.db, self.album, self)
