+++ gui/UserWidget.py
@@ -1,18 +1,20 @@
 from PySide2.QtWidgets import QPushButton
 from PySide2.QtWidgets import QWidget
-from PySide2.QtWidgets import QGridLayout
+from PySide2.QtWidgets import QHBoxLayout
+from PySide2.QtWidgets import QMessageBox
 from gui.EditUserDialog import EditUserDialog
+
 
 class UserWidget(QWidget):
     
     def __init__(self, db, user, parent):
-        super().__init__(parent)
+        super().__init__()
 
         self.db = db 
         self.user = user
         self.parent = parent
 
-        self.layout = QGridLayout()
+        self.layout = QHBoxLayout()
 
         self.init_user()
         self.init_delete()
@@ -23,25 +25,42 @@ class UserWidget(QWidget):
 
     def init_user(self):
         self.user_button = QPushButton(self.user.name)
-        self.layout.addWidget(self.user_button, 0, 0, 1, 2)
+        self.user_button.clicked.connect(self.select)
+        self.layout.addWidget(self.user_button)
+        self.layout.addStretch()
 
         
     def init_delete(self):
         self.delete_button = QPushButton("-")
         self.delete_button.clicked.connect(self.delete)
-        self.layout.addWidget(self.delete_button, 0, 2, 1, 1)
+        self.layout.addWidget(self.delete_button)
 
     def init_edit(self):
         self.edit_button = QPushButton("Edit")
         self.edit_button.clicked.connect(self.edit)
-        self.layout.addWidget(self.edit_button, 0, 3, 1, 1)
+        self.layout.addWidget(self.edit_button)
 
     def delete(self):
-        self.db.delete_user(self.user)
-        self.db.user_view.refresh()
-        self.db.album_view.refresh()
+        msg_box = QMessageBox()
+        msg_box.setWindowTitle("Delete User?")
+        msg_box.setText("Yeet?")
+        msg_box.setStandardButtons(QMessageBox.Yes | QMessageBox.Cancel)
+        msg_box.setDefaultButton(QMessageBox.Cancel)
+        ret = msg_box.exec_()
+
+        if ret == QMessageBox.Yes:    
+            if self.db.current_user == self.user:
+                self.db.current_user = None
+            self.db.delete_user(self.user)
+            self.db.user_view.refresh()
+            self.db.album_view.refresh()
+            self.db.global_rating.refresh()
+            self.db.user_rating.refresh()
 
     def edit(self):
         dialog = EditUserDialog(self.db, self.user, self)
 
-
+    def select(self):
+        self.db.current_user = self.user
+        self.db.user_rating.refresh()
+        self.parent.setSelected(True)
